requirements.txt
streamlit
Pillow
transformers
nltk
textblob
torch